# docs/swagger.yaml
definitions:
  Allowance:
    properties:
      allowanceType:
        enum:
        - donation
        - k-receipt
        type: string
      amount:
        minimum: 0
        type: number
    required:
    - allowanceType
    type: object
  DeductionRequest:
    properties:
      amount:
        type: number
    type: object
  ErrorResponse:
    properties:
      message:
        type: string
    type: object
  PersonalResponse:
    properties:
      personalDeduction:
        type: number
    type: object
  TaxLevel:
    properties:
      level:
        type: string
      tax:
        type: number
    type: object
  TaxRequest:
    properties:
      allowances:
        items:
          $ref: '#/definitions/Allowance'
        type: array
      totalIncome:
        example: 500000
        minimum: 0
        type: number
      wht:
        minimum: 0
        type: number
    type: object
  TaxResponse:
    properties:
      tax:
        type: number
      taxLevel:
        items:
          $ref: '#/definitions/TaxLevel'
        type: array
      taxRefund:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: K-Tax Calculate API
  title: K-Tax API
  version: "1.0"
paths:
  /admin/deductions/personal:
    post:
      consumes:
      - application/json
      description: To setting personal deduction amount for use in tax calculate
      parameters:
      - description: new amount that you want to set
        in: body
        name: tax
        required: true
        schema:
          $ref: '#/definitions/DeductionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PersonalResponse'
        "400":
          description: validate error or cannot get body
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: data not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Personal Deduction Config API
      tags:
      - admin
      - deduction
  /tax/calculations:
    post:
      consumes:
      - application/json
      description: To calculate personal tax and return how much addition pay tax
        / refund tax
      parameters:
      - description: tax data that want to calculate
        in: body
        name: tax
        required: true
        schema:
          $ref: '#/definitions/TaxRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TaxResponse'
        "400":
          description: validate error or cannot get body
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Tax Calculate API
      tags:
      - tax
swagger: "2.0"